+++ R/Senn2013.R
@@ -59,8 +59,8 @@
 #' 
 #' # Fixed effects model
 #' #
-#' net1 <- netmeta(TE, seTE, treat1.long, treat2.long,
-#'                 studlab, data = Senn2013,
+#' net1 <- netmeta(TE, seTE, treat1.long, treat2.long, studlab,
+#'                 data = Senn2013, sm = "MD",
 #'                 comb.random = FALSE, nchar.trts = 4)
 #' net1
 #' net1$Q.decomp